+++ app/controllers/users_controller.rb
@@ -39,11 +39,11 @@ class UsersController < ApplicationController
 
       if current_user.system_admin?
         params.require(:user).permit(
-          :first_name, :last_name, :email, :system_admin
+          :first_name, :last_name, :email, :research_summary, :aug_member, :system_admin
         )
       else
         params.require(:user).permit(
-          :first_name, :last_name, :email
+          :first_name, :last_name, :email, :research_summary
         )
       end
     end