+++ app/controllers/welcome_controller.rb
@@ -4,9 +4,9 @@ class WelcomeController < ApplicationController
 
   def whatsmyip
   end
-  
+
   def aug
-    @users = User.current.where( aug_member: true).order(:last_name).page(params[:page]).per( 40 )
+    @users = User.aug_members.order(:first_name, :last_name).page(params[:page]).per( 40 )
   end
 
   def collection