+++ db/migrate/20140307170527_add_aug_member_and_research_summary_to_users.rb
@@ -1,4 +1,4 @@
-class AddColumnsToUsers < ActiveRecord::Migration
+class AddAugMemberAndResearchSummaryToUsers < ActiveRecord::Migration
   def change
     add_column :users, :aug_member, :boolean, null: false, default: false
     add_column :users, :research_summary, :text